requirements.txt
openai>=1.99.0
feedparser>=6.0.11
python-dotenv>=1.0.1
requests>=2.32.3
